2017-12-11
Made some small changes to the 'voma__stim_analysis' GUI to allow the user to define a 'pre-stim' duration to add to each cycle index. Now, to analyze impulse data, the user can pick a non-zero threshold to trigger each impulse, and choose a time duration to add to each trigger time.