Analyzing Eyes Separately 12/22/17
Selecting to analyze one eye, checking its box with the function on will automatically only plot that eye and will switch when analyzing the next eye.
A popup window comes up when the function is turned on and will automatically close when the function is turned off, or when the cycle analysis gui is closed.
A dialog box pops up to remind the user if an average plot is not made prior to trying to save.
A dialogue box will pop up if go back to qpr gui is clicked prior to both sets of eye data being saved. 
After analyzing and saving data of one eye the number of cycles will reset.
A list of the cycle numbers kept for each eye are saved in the cycle average file and voma file as l_cyclist and r_cyclist when analyzing separately but then just cyclist or cyc2plot for both and also save in voma file. If analyzed separately, there will be no entry for cyc2plot but if analyzed together there will be no entries for l_cyclist and r_cyclist, for the purpose of pulling cycle lists later.