2018-05-22 Created a new function on the 'LDVOG' section of the 'voma__seg_data' routine.
PJB recently found an issue where each LDVOG goggle set delays the recording of the PCU trigger by an integer number of samples. After some experiments determining the approximate sample delay for each goggle set currently in use, PJB added code in the 'voma__seg_data' routine to: correct the trigger delay, save a new .txt file with the raw data, and save the segmented data with a suffix explaining the change that was made.