2018-10-10 In general, this commit is adding two routines to re-calibrate LDVOG data that was collected with an incorrect 'Subject-to-Calibration Grid Distance".
Each routine has a different purpose and is located in a DIFFERENT VOMA GUI:

1) 'voma__seg_data.m'
This routine allows a user to re-calibrate a 'raw' LD VOG file if they have the correct calibration distance and the LEFT and RIGHT eye calibration files.

2) 'voma'
This new routine added to the main VOMA menu is a batch procedure that can be used when an analyst has fully analyzed LD VOG data and realized afterward that the files were collected with the incorrect calibration distance.